Move scientific-notation-scan.ts from language/ to rules/gml/
Previously: lint/src/language/scientific-notation-scan.ts
Now:        lint/src/rules/gml/scientific-notation-scan.ts

Rationale: scientific-notation-scan.ts is a pure text-scanning utility
(a regex pattern + two stateless functions) with no ESLint
language-integration logic. Placing it in language/ forced the
no-scientific-notation lint rule to reach three directory levels up into
the language-integration layer to import it:

  rules/gml/rules/no-scientific-notation-rule.ts
    → ../../../language/scientific-notation-scan.js   (was)
    → ../scientific-notation-scan.js                  (now)

Rules should not depend on the language integration layer's internal
files. Moving the file to rules/gml/ (alongside rule-base-helpers.ts and
rule-helpers.ts) makes it a first-class rule-support utility. The
language/recovery.ts caller is updated to import from the new path
(../rules/gml/scientific-notation-scan.js). No logic was changed; only
the file location and the two import specifiers were updated.
