+++ src/lint/src/rules/gml/scientific-notation-scan.ts
